Resolving mypy error, arg-type error

+++ 0x00-python_variable_annotations/102-type_checking.py
@@ -3,7 +3,7 @@
 from typing import List, Tuple, Union
 
 
-def zoom_array(lst: List, factor: int = 2) -> List[int]:
+def zoom_array(lst: List, factor: float = 2) -> List[int]:
     """Creates multiple copies of items in a tuple, @lst."""
     zoomed_in: List[int] = [
         item for item in lst
